// src/app/models/User.ts
 

export interface User {
    id: string;
    name: string;
    lastName: string;
    username: string;
    email: string;
    Address: string;
    dOB: Date;
    phone: string;
    companyName: string;
    userName: string;
    password: string;
    companyEmail: string;
    companyPhone: string;
    companyAddress: string;
    costalCode: string;
    country: string;
    city: string;
    photoUrl: string;
    language: string;
    userRole: string[];
    Role: string;
    logo: File;
    postalCode:string;
    dateOfBirth:string;
}